split revise+renew out of negotiate
