+++ modules/renter/contractor/negotiate.go
@@ -230,186 +230,3 @@ func (c *Contractor) newContract(host modules.HostSettings, filesize uint64, dur
 
 	return contract, nil
 }
-
-// negotiateRevision sends the revision and new piece data to the host.
-func negotiateRevision(conn net.Conn, rev types.FileContractRevision, piece []byte, secretKey crypto.SecretKey) (types.Transaction, error) {
-	conn.SetDeadline(time.Now().Add(5 * time.Minute)) // sufficient to transfer 4 MB over 100 kbps
-	defer conn.SetDeadline(time.Time{})               // reset timeout after each revision
-
-	// create transaction containing the revision
-	signedTxn := types.Transaction{
-		FileContractRevisions: []types.FileContractRevision{rev},
-		TransactionSignatures: []types.TransactionSignature{{
-			ParentID:       crypto.Hash(rev.ParentID),
-			CoveredFields:  types.CoveredFields{FileContractRevisions: []uint64{0}},
-			PublicKeyIndex: 0, // renter key is always first -- see negotiateContract
-		}},
-	}
-	// sign the transaction
-	encodedSig, _ := crypto.SignHash(signedTxn.SigHash(0), secretKey) // no error possible
-	signedTxn.TransactionSignatures[0].Signature = encodedSig[:]
-
-	// send the transaction
-	if err := encoding.WriteObject(conn, signedTxn); err != nil {
-		return types.Transaction{}, errors.New("couldn't send revision transaction: " + err.Error())
-	}
-
-	// host sends acceptance
-	var response string
-	if err := encoding.ReadObject(conn, &response, 128); err != nil {
-		return types.Transaction{}, errors.New("couldn't read host acceptance: " + err.Error())
-	}
-	if response != modules.AcceptResponse {
-		return types.Transaction{}, errors.New("host rejected revision: " + response)
-	}
-
-	// transfer piece
-	if _, err := conn.Write(piece); err != nil {
-		return types.Transaction{}, errors.New("couldn't transfer piece: " + err.Error())
-	}
-
-	// read txn signed by host
-	var signedHostTxn types.Transaction
-	if err := encoding.ReadObject(conn, &signedHostTxn, types.BlockSizeLimit); err != nil {
-		return types.Transaction{}, errors.New("couldn't read signed revision transaction: " + err.Error())
-	}
-
-	if signedHostTxn.ID() != signedTxn.ID() {
-		return types.Transaction{}, errors.New("host sent bad signed transaction")
-	}
-
-	return signedHostTxn, nil
-}
-
-// newRevision revises the current revision to incorporate new data.
-func newRevision(rev types.FileContractRevision, pieceLen uint64, merkleRoot crypto.Hash, piecePrice types.Currency) types.FileContractRevision {
-	// prevent a negative currency panic
-	if piecePrice.Cmp(rev.NewValidProofOutputs[0].Value) > 0 {
-		// probably not enough money, but the host might accept it anyway
-		piecePrice = rev.NewValidProofOutputs[0].Value
-	}
-	return types.FileContractRevision{
-		ParentID:          rev.ParentID,
-		UnlockConditions:  rev.UnlockConditions,
-		NewRevisionNumber: rev.NewRevisionNumber + 1,
-		NewFileSize:       rev.NewFileSize + pieceLen,
-		NewFileMerkleRoot: merkleRoot,
-		NewWindowStart:    rev.NewWindowStart,
-		NewWindowEnd:      rev.NewWindowEnd,
-		NewValidProofOutputs: []types.SiacoinOutput{
-			// less returned to renter
-			{Value: rev.NewValidProofOutputs[0].Value.Sub(piecePrice), UnlockHash: rev.NewValidProofOutputs[0].UnlockHash},
-			// more given to host
-			{Value: rev.NewValidProofOutputs[1].Value.Add(piecePrice), UnlockHash: rev.NewValidProofOutputs[1].UnlockHash},
-		},
-		NewMissedProofOutputs: []types.SiacoinOutput{
-			// less returned to renter
-			{Value: rev.NewMissedProofOutputs[0].Value.Sub(piecePrice), UnlockHash: rev.NewMissedProofOutputs[0].UnlockHash},
-			// more given to void
-			{Value: rev.NewMissedProofOutputs[1].Value.Add(piecePrice), UnlockHash: rev.NewMissedProofOutputs[1].UnlockHash},
-		},
-		NewUnlockHash: rev.NewUnlockHash,
-	}
-}
-
-// Renew negotiates a new contract for data already stored with a host. It
-// returns the ID of the new contract. This is a blocking call that performs
-// network I/O.
-func (c *Contractor) Renew(fcid types.FileContractID, newEndHeight types.BlockHeight) (types.FileContractID, error) {
-	c.mu.RLock()
-	height := c.blockHeight
-	hc, ok := c.contracts[fcid]
-	c.mu.RUnlock()
-	if !ok {
-		return types.FileContractID{}, errors.New("no record of that contract")
-	} else if newEndHeight < height {
-		return types.FileContractID{}, errors.New("cannot renew below current height")
-	}
-	var host modules.HostSettings
-	found := false
-	for _, h := range c.hdb.AllHosts() {
-		if h.NetAddress == hc.IP {
-			host = h
-			found = true
-			break
-		}
-	}
-	if !found {
-		return types.FileContractID{}, errors.New("no record of that host")
-	} else if host.Price.Cmp(maxPrice) > 0 {
-		return types.FileContractID{}, errTooExpensive
-	}
-
-	// get an address to use for negotiation
-	c.mu.Lock()
-	if c.cachedAddress == (types.UnlockHash{}) {
-		uc, err := c.wallet.NextAddress()
-		if err != nil {
-			c.mu.Unlock()
-			return types.FileContractID{}, err
-		}
-		c.cachedAddress = uc.UnlockHash()
-	}
-	ourAddress := c.cachedAddress
-	c.mu.Unlock()
-
-	renterCost := host.Price.Mul(types.NewCurrency64(hc.LastRevision.NewFileSize)).Mul(types.NewCurrency64(uint64(newEndHeight - height)))
-	renterCost = renterCost.MulFloat(1.05) // extra buffer to guarantee we won't run out of money during revision
-	payout := renterCost                   // no collateral
-
-	// create file contract
-	fc := types.FileContract{
-		FileSize:       hc.LastRevision.NewFileSize,
-		FileMerkleRoot: hc.LastRevision.NewFileMerkleRoot,
-		WindowStart:    newEndHeight,
-		WindowEnd:      newEndHeight + host.WindowSize,
-		Payout:         payout,
-		UnlockHash:     types.UnlockHash{}, // to be filled in by negotiateContract
-		RevisionNumber: 0,
-		ValidProofOutputs: []types.SiacoinOutput{
-			// nothing returned to us; everything goes to the host
-			{Value: types.ZeroCurrency, UnlockHash: ourAddress},
-			{Value: types.PostTax(height, renterCost), UnlockHash: host.UnlockHash},
-		},
-		MissedProofOutputs: []types.SiacoinOutput{
-			// nothing returned to us; everything goes to the void
-			{Value: types.ZeroCurrency, UnlockHash: ourAddress},
-			{Value: types.PostTax(height, renterCost), UnlockHash: types.UnlockHash{}},
-		},
-	}
-
-	// create transaction builder
-	txnBuilder := c.wallet.StartTransaction()
-
-	// initiate connection
-	conn, err := c.dialer.DialTimeout(hc.IP, 15*time.Second)
-	if err != nil {
-		return types.FileContractID{}, err
-	}
-	defer conn.Close()
-	if err := encoding.WriteObject(conn, modules.RPCRenew); err != nil {
-		return types.FileContractID{}, errors.New("couldn't initiate RPC: " + err.Error())
-	}
-	if err := encoding.WriteObject(conn, fcid); err != nil {
-		return types.FileContractID{}, errors.New("couldn't send contract ID: " + err.Error())
-	}
-
-	// execute negotiation protocol
-	newContract, err := negotiateContract(conn, hc.IP, fc, txnBuilder, c.tpool)
-	if err != nil {
-		txnBuilder.Drop() // return unused outputs to wallet
-		return types.FileContractID{}, err
-	}
-
-	// update host contract
-	c.mu.Lock()
-	c.contracts[newContract.ID] = newContract
-	c.cachedAddress = types.UnlockHash{} // clear cachedAddress
-	err = c.save()
-	c.mu.Unlock()
-	if err != nil {
-		c.log.Println("WARN: failed to save the hostdb:", err)
-	}
-
-	return newContract.ID, nil
-}
